Update dependencies (#192)
* Update dependencies

* Fix clippy

+++ datanymizer_dumper/src/postgres/schema_inspector.rs
@@ -74,7 +74,7 @@ impl SchemaInspector for PgSchemaInspector {
                 };
 
                 match self.get_table_size(connection, &table) {
-                    Ok(size) => table.size = size as i64,
+                    Ok(size) => table.size = size,
                     Err(e) => panic!("ERR: {}", e),
                 }
